requirements.txt
openpyxl
pypdf
python-docx
python-pptx
